feat: move tree_sitter to specific file
